rt-39904: CGI::Carp fatalsToBrowser does not catch stringifiable objects
Giving a chance to objects passed to CGI::Carp::die to be stringified, if
they are $WRAPped.

If we are in a non-$WRAP scenario, the object will be left
intact so that the try-catch mechanisms relying on it are not broken.
The first patch for this bug was not taken this behavior in consideration,
and ended up breaking Bugzilla and other applications.

In order not to make the same mistake again, the carp.t tests have been
augmented with as much different ways to die() as possible, to make sure
that we have the desired behavior in all cases.