Refactor: port cleaned-up as_string implementation from CGI::Simple::Cookie
From a careful review, the implementations are identical

One small change relates to this ugly line which was cleaned up:

  my($cookie) = join("=",(defined $key ? $key : ''),join("&",map escape(defined $_ ? $_ : ''),$self->value));

We were doing a couple of ugly checks for definedness to satisfy "use
warnings", which CGI::Simple never enabled. Now I just set "no warnings"
for this block.

Now it's a lot nicer:

    my $value = join "&", map { escape($_) } $self->value;
    my @cookie = ( "$name=$value" );