prep for full release to CPAN
Changes added from recent dev releases. remove the cgi_docs.html
file as this has been deprecated for almost 10 years now and there
has been POD for a long time

the version has been bumped to v4.00 to the make it clear when the
module ceased to be in the perl core. because of this change any
future versions of the module may introduce backwards incompatible
changes and introduce dependencies outside of the perl code, i.e.
other CPAN modules