resolve #111 [rt.cpan.org #82364] - carp timestamp
in CGI::Carp can switched off by use of the import noTimestamp or
by setting the package variable $CGI::Carp::NO_TIMESTAMP to 1. the
alternative solutions to doing this, monkey patching the methods,
are simple to do; but i am taking the path of least resistance in
following the current conventions of these modules by adding the
import switch / package variable

update t/carp.t to test for the timestamp switch both by the use
of the package variable and import option, along with perldoc in
the CGI::Carp module to document the feature.

tidy up some of the perldoc in CGI::Carp and remove the CHANGE LOG
as these are in the Changes file so there is no point duplicating
this information

thanks to EDAVIS@cpan.org for the patch