check #49 [rt.cpan.org #23823] Content-Disposition
filename header does not need to be quoted and will still be
parsed, modify test case to confirm that this is indeed the case.

there are some considerations here:

	- there is a "name" param in this header however there is no
reference to its specification it in any of the following RFCs:

		http://www.ietf.org/rfc/rfc2183.txt
		http://www.ietf.org/rfc/rfc2045.txt
		http://www.ietf.org/rfc/rfc1867.txt
		http://www.ietf.org/rfc/rfc2388.txt

	- given that the "name" param doesn't even exist in RFC 2183 it
is unclear if this is even correct, however it is referenced in the
other RFCs (1867) so erring on the side of caution and leaving it as
is seems to the best course of action (this param must be quoted for
parsing in the current implementation).

update Changes with latest changes to be valid for parsing on CPAN
and bump VERSION for next release.