resolve #55 [rt.cpan.org #36312] - redirect query string
handling, specifically with mod_perl2 - use the ->prev method of
the Apache::RequestUtil object to get at the *previous* request
object and thus its env variables for getting the query string

in theory this would support an unlimited number of redirects as
the next redirect will always pass on the details from the previous
redirect

this is mod_perl2 only, i cannot see this in Apache->request so it
does not appear to be supported in mod_perl1.*. To get around this
use the REDIRECT_ env variable for up to 5 redirects, this also
handles up to 5 redirects when not running under mod perl.

the logic to get the query string from the ENV has been split into
its own method to remove duplication, and a test has been added.
currently there are no tests for mod_perl, Apache, Apache2 use here,
but i don't see any tests at all for any parts of the code using
these modules given i don't have them installed and all the tests
currently pass (the Apache/Apache2 modules are runtime requires
so not hard dependencies)