resolve #60 [rt.cpan.org #50576] - cookie -max-age
argument to CGI::Cookie is now correctly set when this is passed in
the constructor. previously the value from the -expires argument was
being used, meaning it was not possible to pass *only* the -max-age
argument and have this set. this also means that the values for the
-max-age and -expires arguments can be different, should you want to
do something like that

update tests to reflect above change and increase coverage for the
-max-age constructor argument / max_age method. also update perldoc
adding missing documentation for the max_age method