resolve #53 [rt.cpan.org #31107] - test case
for multipart data sending xml content. there is nothing to do here,
if you post multipart form data you MUST include CRLF in the boundary
delimiters as this is part of the protocol. see the following RFC:

	http://www.ietf.org/rfc/rfc2046.txt

note the potentially misleading part in section 5.1.1:

    "NOTE: The CRLF preceding the boundary delimiter line is
	conceptually attached to the boundary so that it is possible
	to have a part that does not end with a CRLF (line break)."

but when one reads the BNF, we notice this isn't always true:

     dash-boundary := "--" boundary
                      ; boundary taken from the value of
                      ; boundary parameter of the
                      ; Content-Type field.

     multipart-body := [preamble CRLF]
                       dash-boundary transport-padding CRLF
                       body-part *encapsulation
                       close-delimiter transport-padding
                       [CRLF epilogue]

     encapsulation := delimiter transport-padding
                      CRLF body-part

     delimiter := CRLF dash-boundary

wherein we see that the first MIME multipart dash-boundary doesn't
include a CRLF.

irrespective of the above, we are to spec and your MIME headers must
include CRLF (well, unless you're on VMS)