resolve #61 [rt.cpan.org #51020] - tied delete returns var
deleted from the hash to make it behave the same as the perl's
built in delete function. add test case to t/request.t for this.

note that this was patch was source from the bugzilla subclass of
CGI (Bugzilla::CGI), which may have more useful fixes as well. see
issue #141 for information on this