failing test case for #62 [rt.cpan.org #51021]
storing an ARRAY ref (or any kind of ref really) in the tied Vars
interface needs to work, but currently the STORE function always
stores an ARRAY ref - the FETCH function returns this by joining
on a null char (\0).

to make this work we will need to somehow retain the way the vars
were passed in originally so we can then choose what the right
thing to return is. the test case added here is currently failing

really i think this whole null char thing needs to go away, it
appears to be some sort of carry over from the perl 4 cgi-lib.pl
days and should be replaced with just DTRT by storing the scalar,
hash, array, whatever. can't do this at present as it will break
back compatibility - something to do in the v5 onwards releases