WIP: #162 - purge all the AUTOLOAD optimisation
but there's a problem - html tag functions are auto generated so we
might be missing a few here. needs more tests, not confident we can
just purge all this code without introducing bugs and/or breaking
back compatibility for some users. also the t/pretty.t test is now
failing (TODO: investigate)

N.B. should this branch ever be merged it should be squashed due to
the failing test in this commit