resolve #137, #162 - split out POD, kill AUTOLOAD
to keep back compatibility we have to auto-generate all the html tag
functions. this change also deprecated CGI::Pretty as it was doing
magic with its own AUTOLOAD function and i'm not interested in the
continued maintenance of this module, alternatives are suggested in
the POD

split out CGI.pm POD into CGI.pod to cut the module size in half so
it's easier to work with

this changes needs significant testing, since the test coverage for
the html tag functions isn't comprehensive (we can delete lots of
the tags from the _all_html_tags method with no test failures)

some users of CGI.pm maybe relying (without realising) on AUTOLOAD
functionality catching bugs in their code, or they may have code
that relies on AUTOLOAD in other ways - if they are then this change
will break their code.