ref #162 - back compatibility functions + test
in removing AUTOLOAD we need to check that the necessary back compat
methods are still available to ensure no (or minimal) breakage. to
do this t/html_functions.t has been added that both uses the
-compile pragma and ->compile function and also checks a function
(a()) is imported into the test's namespace.

the test was run against the version of the module prior to the
removal of AUTOLOAD and then the unnecessary functions were removed
from the _all_html_tags() function in CGI.pm

bump VERSION to dev release, document Changes