resolve #176 - fix handling of PATH_INFO in self_url
w/r/t #126, #152, #176 - according to the CGI spec[1] "No PATH_INFO
segment (see section 4.1.5) is included in the SCRIPT_NAME value" so
we need to make sure this is the case when building up the url. As
this is based from $request_uri or $script_name then we make sure
the PATH_INFO isn't duplicated.

this is complicated by IIS in that it sometimes sets the PATH_INFO
to the *same value* as SCRIPT_NAME, which is a direct contradiction
of the RFC above, so we need to do some special handling for IIS

[1] https://tools.ietf.org/html/rfc3875#section-4.1.13