Add a test that exercises #178
This is in fact more subtle than the original bug report:
$q->tmpFileName($q->param('some_field')) works, because $q->param
for an upload returns a reference that merely behaves like
a filename string. However, stringifying that reference via
"$interpolation", concatenation, or (as in ikiwiki) UTF-8 decoding
gives you a plain string, which used to work but no longer does.