resolve #178 - tmpFileName: fall back if argument is a plain string
In recent versions of CGI.pm, tmpFileName expects a filehandle
(File::Temp reference) argument. In older versions, a plain string
would also work. Try to use the filehandle, but if the argument
is not a reference, fall back to searching for a matching filename.

I haven't done the same for uploadInfo, because uploadInfo was always
documented in terms of a file-handle argument, not a filename.