resolve #196 - sort HTML attributes by default
this was originally done via an undocumented flag in CGI::Util but
is now the default. The HTML spec does not require attributes to be
sorted, but it's useful for testing to get a predictable order back
and also for anything that is watching HTML pages for changes. any
overhead here is moot, this will not be your bottleneck.

bump version and Changes for CPAN release