rename MultipartBuffer package to CGI::MultipartBuffer.
solve some permissions issues and standardise on package names across
the distribution (apart from Fh, which is a back compat thing only)

MultipartBuffer was (is) an undocumented internal package so should
not be in use anywhere, however this has been done in a way to ensure
any $MultipartBuffer package variables are still set correctly in
CGI::MultipartBuffer. if you are explicitly using MultipartBuffer in
a form such as:

	MultipartBuffer->new

your code will break. you should be calling:

	CGI->new->new_MultipartBuffer( $boundary,$length );

to ensure the correctly package is called. if you are extending the
MultipartBuffer package though use of ISA or base (or parent) then you
will need to update your code to use CGI::MultipartBuffer

VERSION has been remove from all but the main CGI package because i'm
sick of having to update umpteen VERSION variables on each release

fake using strict and warnings to appease CPANTS Kwalitee