fix #259 ->url returns the correct thing for ipv6
if a CGI resource is accessed through a IPv6 address in bracket notation
(e.g. http://[::1]:5000), then the url() function handles the host part
wrongly by uri-escaping the brackets. in fact, ->url is uri-escaping
other things that it a) doesn't need to and, b) shouldn't need to

internally replace regexp logic with a call to URI->canonical. the test
coverage around this seems to be pretty good (and poking at the edges
shows more things being fixed that were previously broken).

if any users are uri-unescaping the resultant call to ->url then that
*should* continue to work. there may be edges here however

bump VERSION and Changes for CPAN release