fix #268 - parse quoted cookies with embedded separators
a6b5b931f8f546b4ff887687f0776ad3eacdcaf9 added support for semicolons
*and* commas as the separator for cookies, however there always appears
to have been the bug that quoted (and escaped) cookie values are not
handled correctly:

    #!/usr/bin/env perl

    use strict;
    use warnings;

    use Data::Dumper;
    use CGI::Cookie;

    my %cookies = CGI::Cookie->parse(
        qq{foo=bar; expires="Mon, 07 Apr 2025 21:32:31 GMT"; max-age=7776000}
    );

    print Dumper( [ cookies => \%cookies ] );

gives:

    $VAR1 = [
              'cookies',
              {
                ...
                'expires' => bless( {
                                      'name' => 'expires',
                                      'path' => '/',
                                      'value' => [
                                                   '"Mon'
                                                 ]
                                    }, 'CGI::Cookie' )
              }
            ];

we need to maintain backwards compatibility here, i.e. parsing on both
semicolons and commas, and need to correctly parse quoted values

w/r/t:
 * https://github.com/leejo/CGI.pm/issues/268
 * https://datatracker.ietf.org/doc/html/rfc2109
 * https://datatracker.ietf.org/doc/html/rfc2965
 * https://datatracker.ietf.org/doc/html/rfc6265

RFC 2109 and RFC 2965 say "... A server should also accept comma (,) as
the separator between cookie-values for future compatibility." but then
this note was removed in RFC 6265.

so... messy

replace the split and regex with a char by char parsing to make the
logic clear here. given the max size of a cookie string is 4k this isn't
going to be a bottleneck