fix #271: Revert "fix #270 - handle hash as args with inconsistant dashing"
This reverts commit 24ea55200b9a0a24ebc10116ad1ee03920377a6d.

the change breaks existing documented behaviour:

    $ perl -MCGI -E 'say CGI->VERSION; $q = CGI->new;
        $q->param(test => '-1', '-2', '-3'); say join(", ", $q->multi_param("test"))'
        4.67
        -1, -2, -3

    $ perl -MCGI -E 'say CGI->VERSION; $q = CGI->new;
        $q->param(test => '-1', '-2', '-3'); say join(", ", $q->multi_param("test"))'
        4.68

setting the param like this is documented and this change is not
backwards compatible.

CGI.pm's interface is horrible and broken[^1] but back compatibility is
paramount.

bump VERSION and Changes for CPAN release

[^1]: trying to have an interface like this to a perl subroutine looks
elegant but is fraught with issues on the argument parsing side